fix/ref/chore: bug fixes, code improvements...
There may be more minor changes than are listed here.

update logout to only log the user out in the frontend if the request to log them out in the backend was successful.

change alert when not logged in to a confirm offering to reload the page or log out (authInterceptor)

Add toasts on long loading screens suggesting sth might be wrong, and to reload

(backend) improve isLoggedIn and move it to LoginController

(frontend) improve isLoggedIn, using isFrontAuthenticated and isBackAuthenticated.

Change how CanActivate (authGuardService) checks the credentials (it's now async) due to the above changes.

change/update some comments

fix an issue where loading screen would not drop on activity not found. This modified ActivityService's cache slightly, and coincidentally solved the issue of the activity blacking out on cache expiration (though it may not have happened on this endpoint). A similar approach to this should be used for the mural cache.

add confirmation request on delete activity

fix layout not updating on change/delete/create route (mainly activityController)

center a bunch of modals

fix issue where deleting an activity from a route by clicking the trash icon in the route component gave 500 (securityConfig)

remove "not logged in" alert from profile component

delete {observe: 'response'} from userService.delete() as it's actually not needed. (The code is only accessed in the error, which already receives the status without observe:response)

fix error handling in activityService, and add better logging.

fix error handling in profileComponent

Delete stale code responsible for creating a mural with a list of users. This was dropped in favor of creating an empty mural and letting users join later. Some code remained. It has been deleted.

Hitting cancel on createMural did nothing.

Update labels for current/new password when changing it.

+++ frontend/ATRA/src/app/components/activity/activity.component.ts
@@ -54,7 +54,7 @@ export class ActivityComponent implements OnInit {
         this.receivedAnActivityHandler(act)
         if (act!=null) this.alertService.loaded();
       },
-      error: () => {this.alertService.loaded();console.log("Error: done loading");this.alertService.alert("There was an error fetching the activity. Try reloading the page.")}
+      error: (e) => {this.alertService.loaded();this.alertService.alert("There was an error fetching the activity. Try reloading the page.")}
     })
     this.fetchRoutes()
   }
@@ -72,14 +72,18 @@ export class ActivityComponent implements OnInit {
     this.map.fitBounds(this.path.getBounds());
   }
   deleteActivity() {
-    this.activityService.delete(this.id).subscribe({
-      next: () => {
-        this.alertService.toastSuccess("Route deleted")
-        this.router.navigate(["/me/activities"])
-      },
-      error: (e) => {this.alertService.toastError("There was an error deleting the activity"); console.error("Error deleting activity with id " + this.id + ": "+e.error);
-      },
-    });
+    this.alertService.confirm("This action is irreversible, are you sure you want to continue?", "Deleting activity").subscribe((accepted)=>{
+      if (accepted) {
+        this.activityService.delete(this.id).subscribe({
+          next: () => {
+            this.alertService.toastSuccess("Route deleted")
+            this.router.navigate(["/me/activities"])
+          },
+          error: (e) => {this.alertService.toastError("There was an error deleting the activity"); console.error("Error deleting activity with id " + this.id + ": "+e.error);}
+        });
+      }
+    })
+
   }
 
   //#region route shenanigans
@@ -94,7 +98,7 @@ export class ActivityComponent implements OnInit {
 
   open(content: TemplateRef<any>) {
     if (this.errorLoadingRoutes) return this.alertService.alert("There seem to be no activities with no route assigned.") //check this alongside errorLoadingRoutes
-    this.modal = this.modalService.open(content)
+    this.modal = this.modalService.open(content, { centered:true })
 
     this.selectedRoute = this.activity.route!=null ? this.activity.route.id:-1
     this.routeMap = null;
@@ -147,7 +151,7 @@ export class ActivityComponent implements OnInit {
     this.alertService.loading();
     if (routeId=="-1") {
       this.activityService.removeRoute(this.id).subscribe({
-        next: (activity:any) => {this.alertService.loaded();this.receivedAnActivityHandler(activity, "success at deleting existing connection")},
+        next: (activity:Activity) => {this.alertService.loaded();this.receivedAnActivityHandler(activity, "success at deleting existing connection")},
         error: () => {
           console.error("Error removing the route from the activity");
           this.alertService.loaded()
@@ -157,7 +161,7 @@ export class ActivityComponent implements OnInit {
       return
     }
     this.activityService.addRoute(this.id, parseInt(routeId)).subscribe({
-      next: (activity:any) => {this.alertService.loaded();this.receivedAnActivityHandler(activity, "success at creating new connection")},
+      next: (activity:Activity) => {this.alertService.loaded();this.receivedAnActivityHandler(activity, "success at creating new connection")},
       error: () => {
         console.error("Error changing the avtivity's route");
         this.alertService.loaded()
@@ -204,12 +208,12 @@ export class ActivityComponent implements OnInit {
 
   //#endregion
 
-  receivedAnActivityHandler(act:Activity |null, msg?:string) {
+  receivedAnActivityHandler(act:Activity | null, msg?:string) {
     if (act==null) return;
     if (msg) console.log(msg);
     this.owned = this.authService.user.getValue()?.id==act.user.id;
 
-    this.activity = act; //this.activityService.process1(act); this is now being done in the service
+    this.activity = new Activity(act); //equivalent to activityService.proces1(). Neccessary to use activity methods
     this.stats = this.activity.getOverview()
     this.selectedRoute = this.activity.route!=null ? this.activity.route.id:-1;
     this.currentVisibility = this.activity.visibility
